refactor: move shared edge utilities to language_adapter module
Move edges_shared.rs from devql/ingestion/ to host/language_adapter/,
replacing JsTsArtefact/JsTsDependencyEdge with LanguageArtefact/DependencyEdge.
The old file becomes a re-export shim, and devql ingestion types become
type aliases to the language_adapter canonical types. Widen
CanonicalKindProjection method visibility to pub(crate).
