+++ bitloops/src/host/devql/ingestion/edges_shared.rs
@@ -1,496 +1,1 @@
-use super::*;
-
-// Shared edge-building utilities used by all language edge extractors.
-
-pub(super) struct EdgeCollector<'a> {
-    pub(super) out: &'a mut Vec<JsTsDependencyEdge>,
-    pub(super) seen: &'a mut HashSet<String>,
-}
-
-pub(super) struct SymbolLookup<'a> {
-    pub(super) local_targets: &'a HashMap<String, String>,
-    pub(super) imported_symbol_refs: Option<&'a HashMap<String, String>>,
-}
-
-pub(super) struct EdgeTarget {
-    pub(super) to_target_symbol_fqn: Option<String>,
-    pub(super) to_symbol_ref: Option<String>,
-}
-
-pub(super) struct ReferenceCtx<'a> {
-    pub(super) callables: &'a [JsTsArtefact],
-    pub(super) type_targets: &'a HashMap<String, String>,
-    pub(super) value_targets: &'a HashMap<String, String>,
-    pub(super) imported_symbol_refs: &'a HashMap<String, String>,
-}
-
-pub(super) struct CallCtx<'a> {
-    pub(super) callables: &'a [JsTsArtefact],
-    pub(super) callable_name_to_fqn: &'a HashMap<String, String>,
-    pub(super) imported_symbol_refs: &'a HashMap<String, String>,
-}
-
-pub(super) fn smallest_enclosing_callable(
-    line_no: i32,
-    callables: &[JsTsArtefact],
-) -> Option<JsTsArtefact> {
-    callables
-        .iter()
-        .filter(|c| c.start_line <= line_no && c.end_line >= line_no)
-        .min_by_key(|c| c.end_line - c.start_line)
-        .cloned()
-}
-
-pub(super) fn is_control_keyword(name: &str) -> bool {
-    matches!(
-        name,
-        "if" | "for" | "while" | "switch" | "catch" | "return" | "new" | "typeof"
-    )
-}
-
-pub(super) fn parse_import_clause_symbols(
-    clause: &str,
-    module_ref: &str,
-    imported_symbol_refs: &mut HashMap<String, String>,
-) {
-    let trimmed = clause.trim();
-    if trimmed.is_empty() {
-        return;
-    }
-
-    if let Some((default_part, rest)) = trimmed.split_once(',') {
-        let default_alias = default_part.trim();
-        if !default_alias.is_empty() {
-            imported_symbol_refs
-                .insert(default_alias.to_string(), format!("{module_ref}::default"));
-        }
-        parse_import_clause_symbols(rest, module_ref, imported_symbol_refs);
-        return;
-    }
-
-    if trimmed.starts_with('{') && trimmed.ends_with('}') {
-        let inner = &trimmed[1..trimmed.len() - 1];
-        for part in inner.split(',') {
-            let token = part.trim();
-            if token.is_empty() {
-                continue;
-            }
-            if let Some((orig, alias)) = token.split_once(" as ") {
-                let orig = orig.trim();
-                let alias = alias.trim();
-                if !orig.is_empty() && !alias.is_empty() {
-                    imported_symbol_refs.insert(alias.to_string(), format!("{module_ref}::{orig}"));
-                }
-            } else {
-                imported_symbol_refs.insert(token.to_string(), format!("{module_ref}::{token}"));
-            }
-        }
-        return;
-    }
-
-    if let Some(ns) = trimmed.strip_prefix("* as ") {
-        let ns = ns.trim();
-        if !ns.is_empty() {
-            imported_symbol_refs.insert(ns.to_string(), format!("{module_ref}::*"));
-        }
-        return;
-    }
-
-    imported_symbol_refs.insert(trimmed.to_string(), format!("{module_ref}::default"));
-}
-
-pub(super) fn js_ts_reference_target_maps(
-    artefacts: &[JsTsArtefact],
-) -> (HashMap<String, String>, HashMap<String, String>) {
-    let mut type_targets = HashMap::new();
-    let mut value_targets = HashMap::new();
-
-    for artefact in artefacts {
-        let projected_kind = artefact
-            .canonical_kind
-            .as_deref()
-            .and_then(CanonicalKindProjection::from_str);
-        if projected_kind.is_some_and(|kind| {
-            matches!(
-                kind,
-                CanonicalKindProjection::Interface
-                    | CanonicalKindProjection::Type
-                    | CanonicalKindProjection::Enum
-            )
-        }) || artefact.language_kind == "class_declaration"
-        {
-            type_targets
-                .entry(artefact.name.clone())
-                .or_insert_with(|| artefact.symbol_fqn.clone());
-        }
-        if projected_kind.is_some_and(|kind| {
-            matches!(
-                kind,
-                CanonicalKindProjection::Variable | CanonicalKindProjection::Function
-            )
-        }) || artefact.language_kind == "class_declaration"
-        {
-            value_targets
-                .entry(artefact.name.clone())
-                .or_insert_with(|| artefact.symbol_fqn.clone());
-        }
-    }
-
-    (type_targets, value_targets)
-}
-
-pub(super) fn rust_reference_target_maps(
-    artefacts: &[JsTsArtefact],
-) -> (HashMap<String, String>, HashMap<String, String>) {
-    let mut type_targets = HashMap::new();
-    let mut value_targets = HashMap::new();
-
-    for artefact in artefacts {
-        if matches!(
-            artefact.language_kind.as_str(),
-            "struct_item" | "enum_item" | "trait_item" | "type_item"
-        ) {
-            type_targets
-                .entry(artefact.name.clone())
-                .or_insert_with(|| artefact.symbol_fqn.clone());
-        }
-        if matches!(
-            artefact.language_kind.as_str(),
-            "const_item" | "static_item" | "function_item"
-        ) {
-            value_targets
-                .entry(artefact.name.clone())
-                .or_insert_with(|| artefact.symbol_fqn.clone());
-        }
-    }
-
-    (type_targets, value_targets)
-}
-
-pub(super) fn top_level_export_target_map(artefacts: &[JsTsArtefact]) -> HashMap<String, String> {
-    let mut targets = HashMap::new();
-
-    for artefact in artefacts {
-        if artefact.parent_symbol_fqn.is_some() || artefact.name.is_empty() {
-            continue;
-        }
-
-        if matches!(
-            artefact.language_kind.as_str(),
-            "import_statement" | "use_declaration" | "impl_item"
-        ) {
-            continue;
-        }
-
-        targets
-            .entry(artefact.name.clone())
-            .or_insert_with(|| artefact.symbol_fqn.clone());
-    }
-
-    targets
-}
-
-pub(super) fn syntax_node_trimmed_text(node: tree_sitter::Node, content: &str) -> Option<String> {
-    node.utf8_text(content.as_bytes())
-        .ok()
-        .map(str::trim)
-        .map(str::to_string)
-        .filter(|text| !text.is_empty())
-}
-
-pub(super) fn strip_string_delimiters(text: &str) -> String {
-    text.trim()
-        .trim_matches(|ch| matches!(ch, '"' | '\'' | '`'))
-        .to_string()
-}
-
-pub(super) fn push_reference_edge(
-    col: &mut EdgeCollector,
-    from_symbol_fqn: &str,
-    name: &str,
-    line_no: i32,
-    ref_kind: RefKind,
-    lookup: &SymbolLookup,
-) {
-    let trimmed = name.trim();
-    if trimmed.is_empty() {
-        return;
-    }
-
-    let (to_target_symbol_fqn, to_symbol_ref, resolution) =
-        if let Some(target_fqn) = lookup.local_targets.get(trimmed) {
-            (Some(target_fqn.clone()), None, Resolution::Local)
-        } else if let Some(symbol_ref) = lookup.imported_symbol_refs.and_then(|m| m.get(trimmed)) {
-            (None, Some(symbol_ref.clone()), Resolution::Import)
-        } else {
-            return;
-        };
-
-    let key = format!(
-        "{}|{}|{}|{}|{}|{}",
-        from_symbol_fqn,
-        to_target_symbol_fqn.as_deref().unwrap_or(""),
-        to_symbol_ref.as_deref().unwrap_or(""),
-        line_no,
-        ref_kind.as_str(),
-        resolution.as_str()
-    );
-    if !col.seen.insert(key) {
-        return;
-    }
-
-    col.out.push(JsTsDependencyEdge {
-        edge_kind: EdgeKind::References,
-        from_symbol_fqn: from_symbol_fqn.to_string(),
-        to_target_symbol_fqn,
-        to_symbol_ref,
-        start_line: Some(line_no),
-        end_line: Some(line_no),
-        metadata: EdgeMetadata::reference(ref_kind, resolution),
-    });
-}
-
-pub(super) fn push_extends_edge(
-    col: &mut EdgeCollector,
-    from_symbol_fqn: &str,
-    name: &str,
-    line_no: i32,
-    lookup: &SymbolLookup,
-) {
-    let trimmed = name.trim();
-    if trimmed.is_empty() {
-        return;
-    }
-
-    let (to_target_symbol_fqn, to_symbol_ref, resolution) =
-        if let Some(target_fqn) = lookup.local_targets.get(trimmed) {
-            (Some(target_fqn.clone()), None, Resolution::Local)
-        } else if let Some(symbol_ref) = lookup.imported_symbol_refs.and_then(|m| m.get(trimmed)) {
-            (None, Some(symbol_ref.clone()), Resolution::Import)
-        } else {
-            return;
-        };
-
-    let key = format!(
-        "{}|{}|{}|{}|{}",
-        from_symbol_fqn,
-        to_target_symbol_fqn.as_deref().unwrap_or(""),
-        to_symbol_ref.as_deref().unwrap_or(""),
-        line_no,
-        resolution.as_str()
-    );
-    if !col.seen.insert(key) {
-        return;
-    }
-
-    col.out.push(JsTsDependencyEdge {
-        edge_kind: EdgeKind::Extends,
-        from_symbol_fqn: from_symbol_fqn.to_string(),
-        to_target_symbol_fqn,
-        to_symbol_ref,
-        start_line: Some(line_no),
-        end_line: Some(line_no),
-        metadata: EdgeMetadata::none(),
-    });
-}
-
-pub(super) fn symbol_lookup_name_from_node(
-    node: tree_sitter::Node,
-    content: &str,
-) -> Option<String> {
-    let text = node.utf8_text(content.as_bytes()).ok()?;
-    symbol_lookup_name_from_text(text)
-}
-
-pub(super) fn symbol_lookup_name_from_text(text: &str) -> Option<String> {
-    let mut candidate = text.trim();
-    if candidate.is_empty() {
-        return None;
-    }
-
-    if candidate.starts_with("for<") {
-        candidate = candidate
-            .rsplit_once(' ')
-            .map(|(_, tail)| tail)
-            .unwrap_or(candidate)
-            .trim();
-    }
-
-    candidate = candidate.trim_start_matches('&').trim();
-    candidate = candidate.trim_start_matches("mut ").trim();
-    candidate = candidate.trim_start_matches("dyn ").trim();
-    candidate = candidate.trim_start_matches("impl ").trim();
-    candidate = candidate.split('<').next().unwrap_or(candidate).trim();
-    candidate = candidate.rsplit("::").next().unwrap_or(candidate).trim();
-    candidate = candidate.rsplit('.').next().unwrap_or(candidate).trim();
-    candidate = candidate.trim_matches(|ch: char| matches!(ch, '&' | '(' | ')' | ',' | ';'));
-
-    let first = candidate.chars().next()?;
-    if !(first.is_ascii_alphabetic() || first == '_') {
-        return None;
-    }
-
-    Some(candidate.to_string())
-}
-
-/// Returns true when `node` sits inside an `extends_clause` or `extends_type_clause` in the
-/// JS/TS AST. Used to suppress spurious `references` edges for types that are already captured
-/// by an `extends` edge.
-pub(super) fn js_ts_node_is_in_extends_clause(mut node: tree_sitter::Node) -> bool {
-    loop {
-        let Some(parent) = node.parent() else {
-            return false;
-        };
-        match parent.kind() {
-            "extends_clause" | "extends_type_clause" => return true,
-            // Stop at declaration boundaries — never cross into a different symbol's scope.
-            "class_declaration"
-            | "class_body"
-            | "interface_declaration"
-            | "function_declaration"
-            | "method_definition"
-            | "program" => return false,
-            _ => node = parent,
-        }
-    }
-}
-
-pub(super) fn js_ts_identifier_is_value_reference(node: tree_sitter::Node) -> bool {
-    let Some(parent) = node.parent() else {
-        return true;
-    };
-
-    match parent.kind() {
-        "function_declaration"
-        | "method_definition"
-        | "class_declaration"
-        | "interface_declaration"
-        | "type_alias_declaration"
-        | "required_parameter"
-        | "optional_parameter"
-        | "rest_pattern"
-        | "import_specifier"
-        | "namespace_import"
-        | "pair_pattern"
-        | "shorthand_property_identifier_pattern" => false,
-        "variable_declarator" => !node_matches_parent_field(node, parent, "name"),
-        "call_expression" => !node_matches_parent_field(node, parent, "function"),
-        "new_expression" => !node_matches_parent_field(node, parent, "constructor"),
-        "member_expression" => !node_matches_parent_field(node, parent, "property"),
-        "pair" => !node_matches_parent_field(node, parent, "key"),
-        "property_signature" => !node_matches_parent_field(node, parent, "name"),
-        _ => true,
-    }
-}
-
-pub(super) fn rust_identifier_is_value_reference(node: tree_sitter::Node) -> bool {
-    let Some(parent) = node.parent() else {
-        return true;
-    };
-
-    match parent.kind() {
-        "function_item" | "struct_item" | "trait_item" | "type_item" | "const_item"
-        | "static_item" | "parameter" | "self_parameter" | "use_as_clause" => false,
-        "let_declaration" => !node_matches_parent_field(node, parent, "pattern"),
-        "call_expression" => !node_matches_parent_field(node, parent, "function"),
-        "field_expression" => !node_matches_parent_field(node, parent, "field"),
-        "macro_invocation" => !node_matches_parent_field(node, parent, "macro"),
-        // Suppress e.g. `AppServer` in `AppServer::new(...)`.  The identifier is the path
-        // qualifier of a scoped associated call; the `calls` edge already captures the
-        // dependency, so a redundant `references` edge is not needed.
-        "scoped_identifier" => {
-            let Some(grandparent) = parent.parent() else {
-                return true;
-            };
-            !(grandparent.kind() == "call_expression"
-                && node_matches_parent_field(parent, grandparent, "function"))
-        }
-        _ => true,
-    }
-}
-
-pub(super) fn node_matches_parent_field(
-    node: tree_sitter::Node,
-    parent: tree_sitter::Node,
-    field: &str,
-) -> bool {
-    parent
-        .child_by_field_name(field)
-        .map(|child| same_syntax_node(child, node))
-        .unwrap_or(false)
-}
-
-pub(super) fn same_syntax_node(left: tree_sitter::Node, right: tree_sitter::Node) -> bool {
-    left.kind() == right.kind()
-        && left.start_byte() == right.start_byte()
-        && left.end_byte() == right.end_byte()
-}
-
-#[cfg(test)]
-pub(super) fn line_byte_spans(content: &str) -> Vec<(i32, i32)> {
-    if content.is_empty() {
-        return Vec::new();
-    }
-
-    let mut spans = Vec::new();
-    let mut cursor: i32 = 0;
-    for segment in content.split_inclusive('\n') {
-        let start = cursor;
-        let end = start + segment.len() as i32;
-        spans.push((start, end));
-        cursor = end;
-    }
-    spans
-}
-
-#[cfg(test)]
-pub(super) fn line_start_byte(spans: &[(i32, i32)], line: i32) -> i32 {
-    if line <= 0 {
-        return 0;
-    }
-    spans
-        .get((line - 1) as usize)
-        .map(|(start, _)| *start)
-        .unwrap_or(0)
-}
-
-#[cfg(test)]
-pub(super) fn line_end_byte(spans: &[(i32, i32)], line: i32) -> i32 {
-    if line <= 0 {
-        return 0;
-    }
-    if let Some((_, end)) = spans.get((line - 1) as usize) {
-        return *end;
-    }
-    spans.last().map(|(_, end)| *end).unwrap_or(0)
-}
-
-#[cfg(test)]
-pub(super) fn find_block_end_line(lines: &[&str], start_index: usize) -> Option<i32> {
-    let mut found_open = false;
-    let mut depth = 0i32;
-
-    for (line_idx, line) in lines.iter().enumerate().skip(start_index) {
-        for ch in line.chars() {
-            match ch {
-                '{' => {
-                    found_open = true;
-                    depth += 1;
-                }
-                '}' if found_open => {
-                    depth -= 1;
-                    if depth <= 0 {
-                        return Some((line_idx + 1) as i32);
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-
-    if found_open {
-        Some(lines.len() as i32)
-    } else {
-        None
-    }
-}
+pub(super) use crate::host::language_adapter::edges_shared::*;
